% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neo.R, R/theme_neo_x.R
\docType{data}
\name{inksave_color}
\alias{inksave_color}
\title{theme_neo: A extensible simple but pretty production-ready ggplot theme}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
inksave_color

inksave_color
}
\arguments{
\item{layout}{x lines design integer. It can be 4 to display both major and minor grids from x and y, 3 to display both major and minor x lines, 2 to display only major line, 1 to display only minor line and 0 to display none.}
}
\value{
A ggplot theme object

A ggplot theme object
}
\description{
A simple but pretty ggplot theme

A recommended for time series simple but pretty production ready ggplot theme
}
\examples{
require(ggplot2)
ggplot(mtcars, aes(x=  factor(gear))) +
  geom_bar() +
  geom_text(
    stat="count",
    aes(label = after_stat(count)),
    vjust = -0.5
  ) +
  labs(
    title = "mtcars",
    subtitle = "Gear count"
  ) +
  theme_neo()

# Using layout = 4
ggplot(mtcars, aes(x=  factor(gear))) +
  geom_bar() +
  geom_text(
    stat="count",
    aes(label = after_stat(count)),
    vjust = -0.5
  ) +
  labs(
    title = "mtcars",
    subtitle = "Gear count"
  ) +
  theme_neo(layout = 4)
library(nelsonthemes)
library(ggplot2)

# Converting nhtemp time series object to data frame
nhtemp <- data.frame(nhtemp = nhtemp, year = 1912:1971)

# Time Series/Line chart
ggplot(nhtemp, aes(y = nhtemp, x = year)) +
  geom_line() +
  theme_neo_x() +
  labs(
    title = "Average Yearly Temperatures in New Haven",
    subtitle = "New Haven, Connecticut (1912-1971)",
    x = "Year",
    y = "Temperature"
  )
}
\keyword{datasets}
