% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neo_down.R
\name{neo_down}
\alias{neo_down}
\title{neo_down: Attach plot area to bottom (down)}
\usage{
neo_down(add = 0.25)
}
\arguments{
\item{add}{float value of additive range expansion constants. The default value is 0.25. Recommended parameter value: 0.5.}
}
\value{
A ggplot scale_y_continuous object
}
\description{
This function is useful to attach your plot data to the bottom (x line).
}
\examples{
ggplot(iris, aes(x = Sepal.Length)) +
geom_bar(stat = "count") +
  labs(
    title = "Histogram of Iris dataset",
    subtitle = "Variable: Sepal Length",
    x = "Sepal Length",
    y = "Count"
  ) +
  neo_down()

}
