% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neo.R
\name{theme_neo}
\alias{theme_neo}
\title{A simple but pretty ggplot theme}
\usage{
theme_neo()
}
\value{
A ggplot theme object
}
\description{
A simple but pretty ggplot theme
}
\examples{
require(ggplot2)
ggplot(mtcars, aes(x=  factor(gear))) +
geom_bar() +
  scale_y_continuous(
    expand = expansion(c(0, 0.5))
  ) +
  geom_text(
    stat="count",
    aes(label = after_stat(count)),
    vjust = -0.5
  ) +
  labs(
    title = "mtcars",
    subtitle = "Gear count"
  ) +
  theme_neo()
}
