% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neo.R
\name{theme_neo}
\alias{theme_neo}
\title{theme_neo: A extensible simple but pretty production-ready ggplot theme}
\usage{
theme_neo(layout = 2)
}
\arguments{
\item{layout}{x lines design integer. It can be 4 to display both major and minor grids from x and y, 3 to display both major and minor x lines, 2 to display only major line, 1 to display only minor line and 0 to display none.}
}
\value{
A ggplot theme object
}
\description{
A simple but pretty ggplot theme
}
\examples{
require(ggplot2)
ggplot(mtcars, aes(x=  factor(gear))) +
geom_bar() +
  scale_y_continuous(
    expand = expansion(c(0, 0.5))
  ) +
  geom_text(
    stat="count",
    aes(label = after_stat(count)),
    vjust = -0.5
  ) +
  labs(
    title = "mtcars",
    subtitle = "Gear count"
  ) +
  theme_neo()
}
