% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_neo_x.R
\name{theme_neo_x}
\alias{theme_neo_x}
\title{theme_neo_x: A simple but pretty production ready ggplot theme (x-axis)}
\usage{
theme_neo_x()
}
\value{
A ggplot theme object
}
\description{
A recommended for time series simple but pretty production ready ggplot theme
}
\examples{
require(ggplot2)
ggplot(mtcars, aes(x=  factor(gear))) +
geom_bar() +
  scale_y_continuous(
    expand = expansion(c(0, 0.5))
  ) +
  geom_text(
    stat="count",
    aes(label = after_stat(count)),
    vjust = -0.5
  ) +
  labs(
    title = "mtcars",
    subtitle = "Gear count"
  ) +
  theme_neo_x()
}
